+++ rectools/models/implicit_als.py
@@ -31,8 +31,8 @@ from rectools.exceptions import NotFittedError
 from rectools.models.base import ModelConfig
 from rectools.utils.config import BaseConfig
 from rectools.utils.misc import get_class_or_function_full_path, import_object
+from rectools.utils.serialization import RandomState
 
-from .base import RandomState
 from .rank import Distance
 from .vector import Factors, VectorModel
